Show Month Labels Above & Below Chart
Add 3-letter Month Labels to the top & bottom x-axes, separating them
with continuations of the vertical grid lines.

This adds a new renderMonthLabels diagram & splits out the vertical grid
line generation into the renderMonthLines function so we can vary the
line height & reuse it for the grid & the month labels.