Add Harvesting Dates to Charts
Add new RangeType & RangeBar types to the Types module for specifying if
a bar is for Harvest or Planting and for specifying if a plant's range
is a single continuous range or multiple ranges that may overlap. Add
helper functions to count the number of days in a list of ranges, check
if there are overlaps between two lists of ranges, normalize the ranges
so they do not wrap around the year, & merge identical range types that
are adjacent to each other.

Modify the Parser module to also read the spring & fall harvest dates
from the zone spreadsheets and to build the list of Range Bars for a
Plant. A Plant's range bars will be split into two rows for multiple
reasons, e.g. the planting & harvest dates overlap, there is a large
harvest range, etc.

Modify the Draw module so it renders a single row of bars or multiple
rows of bars depending on how the Plant was parsed. Extract the styling
constants from the rendering functions and make them top-level values
for now. Add a legend to the graphic for indicating which bar colors
correspond to Planting & Harvesting dates. Add a function for rendering
the new RangeBar type

Modify the Main module to change the title of the generated charts from
"Planting Dates" to "Planting & Harvesting Dates".